+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- madsim: Add buggify.
+
 ### Fixed
 
 - etcd: Return error on "lease not found".